# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.CarriageDTO:
    properties:
      code:
        description: |-
          Carriage code (e.g., CAR1, CAR2)
          required: true
          example: CAR1
        type: string
      seats:
        description: |-
          List of seats in this carriage
          required: true
        items:
          $ref: '#/definitions/dto.SeatDTO'
        type: array
    required:
    - code
    - seats
    type: object
...
  dto.CreateTrainRequest:
    properties:
      carriages:
        description: |-
          List of carriages with their seats
          required: true
        items:
          $ref: '#/definitions/dto.CarriageDTO'
        type: array
      class:
        description: |-
          Train class (e.g., Executive, Business, Economy)
          required: true
          example: Executive
        type: string
      name:
        description: |-
          Train name
          required: true
          example: Argo Bromo Anggrek
        type: string
    required:
    - carriages
    - class
    - name
    type: object
...
  dto.Response-any:
    properties:
      code:
        description: |-
          HTTP status code
          example: 200
        type: integer
      data:
        description: Response data
      message:
        description: |-
          Response message
          example: Success
        type: string
    type: object
...
  dto.Response-array_dto_StationResponse:
    properties:
      code:
        description: |-
          HTTP status code
          example: 200
        type: integer
      data:
        description: Response data
        items:
          $ref: '#/definitions/dto.StationResponse'
        type: array
      message:
        description: |-
          Response message
          example: Success
        type: string
    type: object
  dto.Response-array_dto_TrainResponse:
    properties:
      code:
        description: |-
          HTTP status code
          example: 200
        type: integer
      data:
        description: Response data
        items:
          $ref: '#/definitions/dto.TrainResponse'
        type: array
      message:
        description: |-
          Response message
          example: Success
        type: string
    type: object
...
  dto.Response-dto_StationResponse:
    properties:
      code:
        description: |-
          HTTP status code
          example: 200
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/dto.StationResponse'
        description: Response data
      message:
        description: |-
          Response message
          example: Success
        type: string
    type: object
  dto.Response-dto_TrainResponse:
    properties:
      code:
        description: |-
          HTTP status code
          example: 200
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/dto.TrainResponse'
        description: Response data
      message:
        description: |-
          Response message
          example: Success
        type: string
    type: object
...
  dto.SeatDTO:
    properties:
      code:
        description: |-
          Seat code (e.g., A1, B2)
          required: true
          example: A1
        type: string
      is_available:
        description: |-
          Whether the seat is available for booking
          example: true
        type: boolean
    required:
    - code
    type: object
  dto.StationRequest:
    properties:
      city:
        description: |-
          City where the station is located
          required: true
          example: Jakarta
        type: string
      code:
        description: |-
          Station code (2-4 characters)
          required: true
          example: GMB
        maxLength: 4
        minLength: 2
        type: string
      name:
        description: |-
          Station name
          required: true
          example: Gambir Station
        type: string
    required:
    - city
    - code
    - name
    type: object
  dto.StationResponse:
    properties:
      city:
        description: |-
          City where the station is located
          example: Jakarta
        type: string
      code:
        description: |-
          Station code
          example: GMB
        type: string
      id:
        description: |-
          Station unique identifier
          example: 507f1f77bcf86cd799439011
        type: string
      name:
        description: |-
          Station name
          example: Gambir Station
        type: string
    type: object
...
  dto.TrainResponse:
    properties:
      carriages:
        description: List of carriages with their seats
        items:
          $ref: '#/definitions/dto.CarriageDTO'
        type: array
      class:
        description: |-
          Train class
          example: Executive
        type: string
      id:
        description: |-
          Train unique identifier
          example: 507f1f77bcf86cd799439011
        type: string
      name:
        description: |-
          Train name
          example: Argo Bromo Anggrek
        type: string
    type: object
...
  dto.UpdateTrainRequest:
    properties:
      carriages:
        description: |-
          List of carriages with their seats
          required: true
        items:
          $ref: '#/definitions/dto.CarriageDTO'
        type: array
      class:
        description: |-
          Train class (e.g., Executive, Business, Economy)
          required: true
          example: Executive
        type: string
      name:
        description: |-
          Train name
          required: true
          example: Argo Bromo Anggrek
        type: string
    required:
    - carriages
    - class
    - name
    type: object
host: localhost:8080
info:
  contact:
    email: fernandolouis55@gmail.com
    name: Louis Fernando
  description: |-
    This is a comprehensive REST API for Indonesian Railway (KAI) ticket booking system.
    The API provides endpoints for user authentication, station management, train management, schedule search, and booking operations.
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
  title: KAI Backend
  version: "1.0"
paths:
...
  /stations:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all railway stations
      produces:
      - application/json
      responses:
        "200":
          description: Stations retrieved successfully
          schema:
            $ref: '#/definitions/dto.Response-array_dto_StationResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.Response-any'
      summary: Get all stations
      tags:
      - Stations
    post:
      consumes:
      - application/json
      description: Create a new railway station (Admin only)
      parameters:
      - description: Station data
        in: body
        name: station
        required: true
        schema:
          $ref: '#/definitions/dto.StationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Station created successfully
          schema:
            $ref: '#/definitions/dto.Response-dto_StationResponse'
        "400":
          description: Bad request - validation failed
          schema:
            $ref: '#/definitions/dto.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.Response-any'
      security:
      - Bearer: []
      summary: Create a new station
      tags:
      - Stations
  /stations/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a railway station (Admin only)
      parameters:
      - description: Station ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Station deleted successfully
          schema:
            $ref: '#/definitions/dto.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.Response-any'
      security:
      - Bearer: []
      summary: Delete station
      tags:
      - Stations
    get:
      consumes:
      - application/json
      description: Retrieve a specific station by its ID
      parameters:
      - description: Station ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Station retrieved successfully
          schema:
            $ref: '#/definitions/dto.Response-dto_StationResponse'
        "404":
          description: Station not found
          schema:
            $ref: '#/definitions/dto.Response-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.Response-any'
      summary: Get station by ID
      tags:
      - Stations
    put:
      consumes:
      - application/json
      description: Update an existing railway station (Admin only)
      parameters:
      - description: Station ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated station data
        in: body
        name: station
        required: true
        schema:
          $ref: '#/definitions/dto.StationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Station updated successfully
          schema:
            $ref: '#/definitions/dto.Response-dto_StationResponse'
        "400":
          description: Bad request - validation failed
          schema:
            $ref: '#/definitions/dto.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.Response-any'
      security:
      - Bearer: []
      summary: Update station
      tags:
      - Stations
  /trains:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all trains
      produces:
      - application/json
      responses:
        "200":
          description: Trains retrieved successfully
          schema:
            $ref: '#/definitions/dto.Response-array_dto_TrainResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.Response-any'
      summary: Get all trains
      tags:
      - Trains
    post:
      consumes:
      - application/json
      description: Create a new train with carriages and seats (Admin only)
      parameters:
      - description: Train data
        in: body
        name: train
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTrainRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Train created successfully
          schema:
            $ref: '#/definitions/dto.Response-dto_TrainResponse'
        "400":
          description: Bad request - validation failed
          schema:
            $ref: '#/definitions/dto.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.Response-any'
      security:
      - Bearer: []
      summary: Create a new train
      tags:
      - Trains
  /trains/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a train (Admin only)
      parameters:
      - description: Train ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Train deleted successfully
          schema:
            $ref: '#/definitions/dto.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.Response-any'
      security:
      - Bearer: []
      summary: Delete train
      tags:
      - Trains
    get:
      consumes:
      - application/json
      description: Retrieve a specific train by its ID
      parameters:
      - description: Train ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Train retrieved successfully
          schema:
            $ref: '#/definitions/dto.Response-dto_TrainResponse'
        "404":
          description: Train not found
          schema:
            $ref: '#/definitions/dto.Response-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.Response-any'
      summary: Get train by ID
      tags:
      - Trains
    put:
      consumes:
      - application/json
      description: Update an existing train (Admin only)
      parameters:
      - description: Train ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated train data
        in: body
        name: train
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTrainRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Train updated successfully
          schema:
            $ref: '#/definitions/dto.Response-dto_TrainResponse'
        "400":
          description: Bad request - validation failed
          schema:
            $ref: '#/definitions/dto.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.Response-any'
      security:
      - Bearer: []
      summary: Update train
      tags:
      - Trains
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: 'JWT Authorization header using the Bearer scheme. Example: "Authorization:
      Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"